% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlabasterParam.R
\name{AlabasterParam}
\alias{AlabasterParam}
\alias{AlabasterParam-class}
\title{Store MS data objects using the alabaster framework}
\usage{
AlabasterParam(path = tempdir())
}
\arguments{
\item{path}{\code{character(1)} with the name of the directory where the MS data
object should be saved to or from which it should be restored.
Importantly, path should point to a \strong{new} folder, i.e. a directory
that \strong{does not already exist}.}
}
\value{
For \code{AlabasterParam()}: an instance of \code{AlabasterParam} class. For
\code{readObject()} the exported object in the specified path (depending on
the type of object defined in the \emph{OBJECT} file in the path. For
\code{readMsObject()} the exported data object, defined with the function's
first parameter, from the specified path. \code{saveObject()} and
\code{saveMsObject()} don't return anything.
}
\description{
The \href{https://github.com/ArtifactDB/alabaster.base}{\emph{alabaster} framework}
provides the methodology to save R objects to on-disk representations/
storage modes which are programming language independent (in contrast to
e.g. R's RDS files). By using standard file formats such as JSON and HDF5,
alabaster ensures that the data can also be read and imported by other
programming languages such as Python or Javascript. This improves
interoperability between application ecosystems.

The \emph{alabaster} package defines the \code{\link[alabaster.base:saveObject]{alabaster.base::saveObject()}} and
\code{\link[alabaster.base:readObject]{alabaster.base::readObject()}} methods that have to be implemented for
specific data classes to enable saving to or reading from alabaster-based
file formats.

In addition, the \emph{MsStash} package defines the \code{AlabasterParam} which can be
used to write or read MS objects using the \code{saveMsObject()} and
\code{readMsObject()} methods in alabaster format. This allows additional
configurations and customizations to the export or import process. It is
thus for example possible to specify the path to the original MS data files
for \emph{on-disk} MS representations such as the \code{Spectra::MsBackendMzR} which
enables to import a stored object even if either the object or the original
MS data files have been moved to a different directory or file system.

In order to enable alabaster-based import/export, the respective methods have
to be implemented for the class. See the package vignette for examples and
details.
}
\details{
Importantly, it is only possible to save \strong{one object in one directory}. To
overwrite an existing stored object in a folder, that folder has to be
deleted beforehand.
}
\examples{

## Create an AlabasterParam object
a <- AlabasterParam(path = tempdir())
a

## See the package vignette for example implementation and usage.
}
\seealso{
Other MS object export and import formats.:
\code{\link[=PlainTextParam]{PlainTextParam()}}
}
\author{
Johannes Rainer, Philippine Louail
}
\concept{MS object export and import formats.}
