% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlainTextParam.R
\name{PlainTextParam}
\alias{PlainTextParam}
\alias{PlainTextParam-class}
\title{Store contents of MS objects as plain text files}
\usage{
PlainTextParam(path = tempdir())
}
\arguments{
\item{path}{For \code{PlainTextParam()}: \code{character(1)}, defining where the files
are going to be stored/ should be loaded from. The default is
\code{path = tempdir()}.}
}
\value{
For \code{PlainTextParam()}: a \code{PlainTextParam} class. \code{saveMsObject()}
does not return anything but saves the object to collections of different
plain text files to a folder. The \code{readMsObject()} method returns the
restored data as an instance of the class specified with parameter \code{object}.
}
\description{
The \code{saveMsObject()} and \code{readMsObject()} methods with the \code{PlainTextParam}
option enable users to save/load different type of mass spectrometry (MS)
object as a collections of plain text files in/from a specified folder.
This folder, defined with the \code{path} parameter, will be created by the
\code{saveMsObject()} function. Writing data to a folder that contains already
exported data will result in an error.

For \code{PlainTextParam} all data is expected to be exported to plain text files,
where possible as tabulator delimited text files.

To support writing/reading with \code{PlainTextParam}, the \code{saveMsData()} and
\code{readMsData()} methods have to be implemented for the respective class.

See the package vignette for example implementations and details.
}
\examples{

## Create a PlainTextParam object
p <- PlainTextParam()
p

## For example implementations and details see the package vignette
}
\seealso{
Other MS object export and import formats.:
\code{\link[=AlabasterParam]{AlabasterParam()}}
}
\author{
Philippine Louail, Johannes Rainer
}
\concept{MS object export and import formats.}
